tests/wpt/web-platform-tests/dom/events/EventTarget-removeEventListener.html

<!DOCTYPE html>
<meta charset="utf-8">
<title>EventTarget.removeEventListener</title>
<link rel="author" title="Ms2ger" href="mailto:Ms2ger@gmail.com">
<link rel="help" href="https://dom.spec.whatwg.org/#dom-eventtarget-removeeventlistener">
<script src="/resources/testharness.js"></script>
<script src="/resources/testharnessreport.js"></script>
<div id="log"></div>
<script>
// Step 1.
test(function() {
  assert_equals(document.removeEventListener("x", null, false), undefined);
  assert_equals(document.removeEventListener("x", null, true), undefined);
  assert_equals(document.removeEventListener("x", null), undefined);
}, "removing a null event listener should succeed");
</script>
